Pypiper can also remove empty directories

+++ pypiper/pypiper.py
@@ -772,7 +772,8 @@ class Pypiper:
 				files = glob.glob(regex)
 				for file in files:
 					with open(self.cleanup_file, "a") as myfile:
-						myfile.write("rm " + file + "\n")
+						if os.path.isfile(file): myfile.write("rm " + file + "\n")
+						elif os.path.isdir(file): myfile.write("rmdir " + file + "\n")
 			except:
 				pass
 		elif conditional:
@@ -799,8 +800,12 @@ class Pypiper:
 					while files in self.cleanup_list: self.cleanup_list.remove(files)
 					# and delete the files
 					for file in files:
-						print("`rm " + file + "`")
-						os.remove(os.path.join(file))
+						if os.path.isfile(file):
+							print("`rm " + file + "`")
+							os.remove(os.path.join(file))
+						elif os.path.isdir(file):
+							print("`rmdir " + file + "`")
+							os.rmdir(os.path.join(file))
 				except:
 					pass
 
@@ -815,8 +820,12 @@ class Pypiper:
 						files = glob.glob(expr)
 						while files in self.cleanup_list_conditional: self.cleanup_list_conditional.remove(files)
 						for file in files:
-							print("rm " + file)
-							os.remove(os.path.join(file))
+							if os.path.isfile(file):
+								print("`rm " + file + "`")
+								os.remove(os.path.join(file))
+							elif os.path.isdir(file):
+								print("`rmdir " + file + "`")
+								os.rmdir(os.path.join(file))
 					except:
 						pass
 			else:
@@ -828,7 +837,8 @@ class Pypiper:
 						files = glob.glob(expr)
 						for file in files:
 							with open(self.cleanup_file, "a") as myfile:
-								myfile.write("`rm " + file + "`\n")
+								if os.path.isfile(file): myfile.write("`rm " + file + "`\n")
+								elif os.path.isdir(file): myfile.write("`rmdir " + file + "`\n")
 					except:
 						pass
